pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bsort"
version = "0.1.0"
description = "Bottlecap Sorter - ML Pipeline"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Dhafindra Stefano", email = "dhafindra.razaqa@gmail.com"}
]
dependencies = [
    "opencv-python",
    "numpy",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "wandb",
    "click",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "flake8",
    "black",
    "isort",
    "pylint",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.pylint.messages_control]
disable = [
    "C0114", # missing-module-docstring
    "C0115", # missing-class-docstring
    "C0116", # missing-function-docstring
    "R0903", # too-few-public-methods
]

[project.scripts]
bsort = "bsort.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
